requirements.txt
##Python: 3.9.9
Flask==2.0.2
##Bootstrap: 5.1.x
##MySQL: 8.0.28
mysql-connector-python==8.0.28
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
Flask-Bcrypt==0.7.1
Werkzeug==2.0.0